restapi/requirements.txt
Django==1.8.4
gunicorn
PyYAML
docker-py
dockerpty
docopt
requests
six
texttable
websocket-client
django_forms_bootstrap
dj_static
django-oauth-toolkit
django-cors-headers
django-allauth
Markdown==2.6.2
django-filter==0.11.0
django-extensions==1.6.1
virtualenv==13.1.0
psycopg2==2.4.5
django-rest-swagger==0.3.4
djangorestframework==3.2.3
djangorestframework-jwt==1.7.2
drf-nested-routers==0.10.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
djrill==2.0
Pillow==2.3
django-imagekit==3.3